fix the nginx integration test
We found the default behaviour of scope to reconnect to tcp:9109 for
events was impacting operation of Nginx. Not sure how - another ticket
coming for that. In the mean time, we've adjusted the env vars to
disable events and set the event destination to a file instead of a
socket.

We've also unset LD_PRELOAD when running `ab` since the intention of
this is to test scoped Nginx, not Apache Benchmark.

Closes #450