passing unit test now
The test that compared the default config to the results of loading our
stock scope.yml was failing because some values should default to null
instead of empty strings. Commented out those entries in the config and
we get null for them now.

Also, by default, HTTP events from the library should be turned on.
Fixed another leftover from the HTTP work to correct this.