comment configs and correct HTTP watch default
Closes #466

Added copious comments to `conf/scope.yml` and adjusted
`conf/scope_protocol.yml` to match. The idea is to document the config
schema there in the stock configs.

During the documentation effort above, found the enable setting for the
HTTP watch was incorrect in the code. The corresponding entries in the
stock config were commented out. This change puts things back to the way
they're supposed to default with the HTTP watch enabled by default and
automatically disabled when connected to LogStream.