removing leftovers from previous work
The azure-pipelines.yml and ContainerReadme.md files should he been
removed when we were adjusting the itegration testing scheme.

The package-lock.json file appears to be an accidental commit as we
don't use NodeJS anywhere in the project.

The other scripts were obsoleted by recent adjustments to the CI
workflows and build process.