(543) fixes after testing browser and server
The HPACK decoder is supposed to be used by all streams on the channel
instead of using a separate one for each stream. I read the spec wrong.

The nghttp2.org site I was testing the browser with is responding to
requests with frames in separate payloads and I found my "stash" logic
was busted. Was copying the stash onto the stack instead of pointing to
it. Fixed.

The decoder defaults to limiting the dynamic table to 4096 entries but
the nghttp2.org site is responding with requests tho increase the table
beyond that resulting in decoder errors. Increased the limit to 16x and
things are working now.

Added some details to log messages for protocol-detect and posted
payloads to help confirm the HTTP handling logic.