(543) apply watch filters to HTTP/2 events
This implements the filters for the HTTP watch. The `name` filter
blocks the events when it doesn't match. The `field` and `value` filters
control which fields appear on the event's body.data object.

Updated docs/HTTP.md and conf/scope.yml to correct the descriptions for
how these filters are applied.

Had to add the `ctlEvtGet()` accessor to get the watch filters.