(543) clarify comments
Neglected to mention this in an earlier commit but the logic for
handling encrypted payloads has changed. See `doProtocol()`.

In the past, when connected to LogStream, we would forward encrypted
payloads until we saw the first unencrypted payload on the channel. This
effectively forwarded TLS handshake payloads then stopped.

We've also been required to prevent  encrypted payloads from being
written to files when not connected to LogStream. This required remains.

What's changed is the forwarding of TLS handshake payloads. We are no
longer doing this at all. We never send or store any encrypted payloads.
The first reason for the planned removal of HTTP parsing from LogStream
so the TLS and HTTP payloads no longer need to be sent. The second
reason is that in some situations where we don't get a file descriptor
for TLS/SSL payloads, we don't have a way to detect when the handshake
is complete so we can't turn them off.