trying to resolve ambiguity in workflow on.push
Without the on.push.tags entry, this workflow is not being triggered
when we create the `web*` tags. Docs at [1] are not terribly clear and
this [post] doesn't seem to clear it up so I'm testing here.

I'm adding the on.push.tags entry in the hope that we can trigger the
website deployment with `web*` tags as we planned. Theory is that the
on.push.branches and on.push.tags entries are OR'ed, not AND'ed.

[1]: https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#onpushpull_requestbranchestags
[post]: https://github.community/t/combine-path-and-tags-conditionals/17064