testing the website workflow
This is a benign change in the website/ directory intended to trigger
the website to be deployed to staging. I'm testing the recent change in
the website workflow to ensure that adding the on.push.tags entry
doesn't block the staging commits for pushed to master without a web*
tag.