(543) case-insensitive header searching
Switched from the original substring search routine to regexes when
looking for headers in HTTP/1 messages to make them case-insensitive as
the HTTP specs say they should be. Before the HTTP/2 additions, we were
only looking for Content-Length on the data side and everythign was
working in our experience. When I used the same matching scheme to look
for the HTTP1->2 upgrade headers, I ran into different cases from
different programs so we needed to make this change.