(543) fix unit test
When we switched from substring searching to a regex for HTTP/1 headers,
we needed `pcre2_match_wrapper` to link httpstatetest. Adding the object
where its defined led to a cascade of other undefined symbols and the
need to wrap `cmdPostEven()`. This was messy but the tests build and
pass now.