(543) workaround protocol-detection issues
When we based the updated HTTP parsing logic on HTTP protocol detection,
we exposed some issues in that detection logic that result from issues
in our TLS/SSL interpositions. These changes address work around some of
those issues. See #600 for some of the detauils as well as the comments
added in this commit.