Ignore new `web*` tags for VERSION
Logic in `Makefile` will extract the version number from git it's not passed in via the `VERSION` envvar. The logic is currently looking for the first tag seen going up the branch. With the addition of the new `web*` tags, the `VERSION` is wrong. This change adjusts the logic to only look for `v*` tags.