Add handling implicit close of `newfd` in `dup2`/`dup3`
While waiting for the actual TCP connection (when connection is
pending) the socket was already created by the AppScope and
stored in `net.pending_connect` field.
The file descriptor returned by the `socket` function will be
the lowest-number file descriptor not currently open for the
process[1] e.g. with value equals 3.

The scoped application can try to use low id file descriptor for
`dup2(oldf,newfd)` function. If the new file descriptor (`newfd`)
will be equal to file descriptor previously used for socket
creation, then the socket will be silenty closed [2] and file
descriptor will be reused for the application purpose:
duplication of `oldfd`.

After this operation value stored in `net.pending_connect` does
not point longer to a socket initially created by the AppScope.
The file descriptor now belongs to the scoped application.

Later AppScope in `checkPendingSocketStatus` will try to verify
socket status using `getsockopt` and can fail if the file
descriptor now points for something different than the socket,
e.g. the pipe. After failing AppScope will close the file
descriptor which now belongs to the scoped application.

This results in an issue that if the application later will try
to use the file descriptor closed by the AppScope it will fail.

 -- <AppScope> Starts TCP connection by creating a socket
 -- <FD state> File descriptor 3 points to a socket
 -- <AppScope> value 3 is stored in `net.pending_connect` field
 -- ...
 -- <Application> Duplicate stdout to fd=3, dup2(1,3)
 -- <FD state> File descriptor 3 points to a pipe
 -- <AppScope> Tries to finish starting connection procedure
 -- <AppScope> `getsockopt` fails
 -- <AppScope> `close` is called for value stored in `net.pending_connect`
 -- <FD state> File descriptor 3 is closed now
 -- ...
 -- <Application> Duplicate fd=3 to stdout, dup2(3,1) fails
 -- <Applicatton> errno=EBADF because the file descriptor 3 was closed

The solution is to close file descriptor owned by the AppScope simillar
as just in case of `close` operation (criblo/appscope#443)

Link: https://man7.org/linux/man-pages/man2/socket.2.html [1]
Link: https://man7.org/linux/man-pages/man2/dup.2.html [2]

closes #657

Co-authored-by: michalbiesek <mbiesek@cribl.io>